refactor: moved tokens into their own crate and refactored entire source code
