+++ crates/cyrusc_ast/Cargo.toml
@@ -5,3 +5,4 @@ edition = "2024"
 
 [dependencies]
 cyrusc_abi = { path = "../cyrusc_abi", version = "*" }
+cyrusc_tokens = { path = "../cyrusc_tokens", version = "*" }
