+++ crates/cyrusc_cir/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+cyrusc_tokens = { path = "../cyrusc_tokens", version = "*" }
 cyrusc_fs_utils = { path = "../cyrusc_fs_utils", version = "*" }
 cyrusc_sema = { path = "../cyrusc_sema", version = "*" }
 cyrusc_tast = { path = "../cyrusc_tast", version = "*" }
