+++ crates/cyrusc_lexer/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+cyrusc_tokens = { path = "../cyrusc_tokens", version = "*" }
 cyrusc_strescape = { path = "../cyrusc_strescape", version = "*" }
 cyrusc_fs_utils = { path = "../cyrusc_fs_utils", version = "*" }
 cyrusc_ast = { path = "../cyrusc_ast", version = "*" }
