+++ crates/cyrusc_parser/src/diagnostics.rs
@@ -14,8 +14,8 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
-use cyrusc_ast::token::TokenKind;
 use cyrusc_diagcentral::DiagKind;
+use cyrusc_tokens::TokenKind;
 use thiserror::Error;
 
 #[derive(Debug, Error, Clone)]
