+++ crates/cyrusc_resolver/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+cyrusc_tokens = { path = "../cyrusc_tokens", version = "*" }
 cyrusc_abi = { path = "../cyrusc_abi", version = "*" }
 cyrusc_modulefsloader = { path = "../cyrusc_modulefsloader", version = "*" }
 cyrusc_diagcentral = { path = "../cyrusc_diagcentral", version = "*" }
