+++ crates/cyrusc_sema/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+cyrusc_tokens = { path = "../cyrusc_tokens", version = "*" }
 cyrusc_abi = { path = "../cyrusc_abi", version = "*" }
 cyrusc_fs_utils = { path = "../cyrusc_fs_utils", version = "*" }
 cyrusc_tast = { path = "../cyrusc_tast", version = "*" }
