+++ crates/cyrusc_sema/src/const_fold.rs
@@ -1,29 +1,27 @@
-/* 
+/*
  * Copyright (c) 2026 The Cyrus Language
- * 
+ *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. 
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
  * See the GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 use crate::analyze::AnalysisContext;
-use cyrusc_ast::{
-    LiteralKind,
-    operators::{InfixOperator, PrefixOperator},
-};
+use cyrusc_ast::operators::{InfixOperator, PrefixOperator};
 use cyrusc_resolver::symbols::{LocalOrGlobalSymbol, LocalScopeRef};
 use cyrusc_tast::{
     ScopeID, SymbolID,
     exprs::{TypedExprKind, TypedExprStmt, TypedLiteralExpr},
 };
+use cyrusc_tokens::literals::LiteralKind;
 
 // TODO: Move entire crate and it's logic to cyrusc_interp.
