+++ crates/cyrusc_tast/src/format.rs
@@ -23,7 +23,8 @@ use crate::{
         TypedUStructType,
     },
 };
-use cyrusc_ast::{AssignmentKind, LiteralKind, StringPrefix, operators::UnaryOperator};
+use cyrusc_ast::{AssignmentKind, operators::UnaryOperator};
+use cyrusc_tokens::literals::{LiteralKind, StringPrefix};
 
 pub fn format_typed_exprs<'a>(exprs: &Vec<TypedExprStmt>, format_symbol: &(dyn Fn(SymbolID) -> String + 'a)) -> String {
     exprs
