+++ crates/cyrusc_tokens/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "cyrusc_tokens"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
